refactor: eliminate code duplication in OpenAI provider
Consolidate Chat and ChatStream implementations by:
- Moving chatWithMessages() and chatStreamWithMessages() from openai_multimodal.go to openai.go
- Moving helper functions (extractContentString, extractTextFromParts, getTextFromPart) to openai.go
- Updating Chat() and ChatStream() to delegate to common implementations
- Removing duplicate implementations from openai_multimodal.go

Results:
- Reduced openai_multimodal.go from 433 to 195 lines (-238 lines, -55%)
- Eliminated ~200 lines of duplicated HTTP request/response handling code
- No functionality changes - all tests still passing
- Significantly reduced code duplication metrics for Sonar

This resolves the major duplication issue flagged by SonarQube:
- openai_multimodal.go: 22.6% → Expected <5%
- openai.go: 19.0% → Expected <5%
